requirements.txt
absl-py
altair
asttokens
astunparse
attrs
backcall
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
colorama
comm
contourpy
cycler
debugpy
decorator
entrypoints
executing
Faker
favicon
flatbuffers
fonttools
gast
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
htbuilder
idna
importlib_metadata
importlib_resources
ipykernel
ipython
jedi
Jinja2
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
keras
kiwisolver
libclang
lxml
Markdown
markdown-it-py
markdownlit
MarkupSafe
matplotlib
matplotlib-inline
mdurl
more-itertools
nest_asyncio
numpy
oauthlib
opt-einsum
packaging
pandas
parso
pickleshare
pillow
pkgutil_resolve_name
platformdirs
prometheus_client
prompt_toolkit
protobuf
psutil
pure_eval
pyarrow
pyasn1
pyasn1_modules
pydeck
Pygments
pymdown-extensions
pyparsing
python-dateutil
pytz
pywin32
PyYAML
pyzmq
referencing
requests
requests-oauthlib
rich
rpds-py
rsa
scikit-learn
scipy
seaborn
six
smmap
soupsieve
st-annotated-text
st-theme
stack-data
streamlit
streamlit-camera-input-live
streamlit-card
streamlit-embedcode
streamlit-extras
streamlit-faker
streamlit-image-coordinates
streamlit-keyup
streamlit-lottie
streamlit-option-menu
streamlit-toggle-switch
streamlit-vertical-slider
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
toml
toolz
tornado
traitlets
typing_extensions
tzdata
urllib3
validators
watchdog
wcwidth
Werkzeug
wrapt
xgboost
zipp
